Simplify Video selection usage
The list of videos was previously opened through a modal window
(using an intent 'PICK' to Drive). It was displayed when the
user clicked on the button "Open video..." which was present
both on the Player page and the Library page.

Now the list is only displayed on the Library page and is shown
directly to the user (without the need to push a button)